refactor: TokensページをplaygroundからStorybookへ移動
TokensページはCSS変数の静的なリファレンス表示であり、SSR/CSR固有の
動作確認という playground の役割に合致しないため、コンポーネントの
見た目を確認する場である Storybook（Foundations/Tokens）に移動した。
