--- playground/nuxt4/app/pages/tokens.vue
@@ -1,4 +1,0 @@
-<script lang="ts">
-import Page from 'playground-shared/pages/TokensPage.vue';
-export default Page;
-</script>
