+++ playground/vue/src/main.ts
@@ -20,8 +20,7 @@ const router = createRouter({
         { path: '/', component: HomePage },
         { path: '/forms', component: () => import('playground-shared/pages/FormsPage.vue') },
         { path: '/feedback', component: () => import('playground-shared/pages/FeedbackPage.vue') },
-        { path: '/navigation', component: () => import('playground-shared/pages/NavigationPage.vue') },
-        { path: '/tokens', component: () => import('playground-shared/pages/TokensPage.vue') }
+        { path: '/navigation', component: () => import('playground-shared/pages/NavigationPage.vue') }
     ]
 });
